Refactor bucket file share dialog
