+++ webui/src/features/bucket/index.ts
@@ -1,1 +1,2 @@
 export {CreateBucketDialog} from "./ui/create-bucket-dialog";
+export {ShareFileDialog} from "./ui/share-file-dialog";
